secure views on admin

+++ vehicles/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from . models import Vehicle, Category
 from .forms import VehicleForm
@@ -67,9 +68,13 @@ def vehicle_detail(request, vehicle_id):
 
     return render(request, 'vehicles/vehicle_detail.html', context)
 
-
+@login_required
 def add_vehicle(request):
     """ Add a vehicle """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only vehicle owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = VehicleForm(request.POST, request.FILES)
         if form.is_valid():
@@ -88,8 +93,13 @@ def add_vehicle(request):
 
     return render(request, template, context)
 
+@login_required
 def edit_vehicle(request, vehicle_id):
     """edit a vehicle """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only vehicle owners can do that.')
+        return redirect(reverse('home'))
+
     vehicle = get_object_or_404(Vehicle, pk=vehicle_id)
     if request.method == 'POST':
         form = VehicleForm(request.POST, request.FILES, instance=vehicle)
@@ -111,8 +121,13 @@ def edit_vehicle(request, vehicle_id):
 
     return render(request, template, context)
 
+@login_required
 def delete_vehicle(request, vehicle_id):
     """delete vehicle"""
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only vehicle owners can do that.')
+        return redirect(reverse('home'))
+        
     vehicle = get_object_or_404(Vehicle, pk=vehicle_id)
     vehicle.delete()
     messages.success(request, 'Vehicle deleted!')
